+++ src/const.js
@@ -1,3 +1,37 @@
+const HIDDEN_CLASS = `visually-hidden`;
+
+const ITEM_ACTIVE_CLASS = `film-card__controls-item--active`;
+
+const FILTER_ID_PREFIX = `filter__`;
+
+const DefaultData = {
+  deleteButtonText: `Delete`,
+};
+
+const SHOWING_CARDS_COUNT_ON_START = 5;
+const SHOWING_CARDS_COUNT_BY_BUTTON = 5;
+const EXTRA_FILM_CARDS_COUNT = 2;
+
+const SHAKE_ANIMATION_TIMEOUT = 600;
+const siteBodyElement = document.querySelector(`body`);
+
+const Mode = {
+  DEFAULT: `default`,
+  OPEN_POPUP: `openPopup`,
+};
+
+const RenderPosition = {
+  AFTERBEGIN: `afterbegin`,
+  BEFOREEND: `beforeend`
+};
+
+const Method = {
+  GET: `GET`,
+  POST: `POST`,
+  PUT: `PUT`,
+  DELETE: `DELETE`
+};
+
 const MONTHS = [
   `January`,
   `February`,
@@ -20,77 +54,6 @@ const COMMENT_EMOJIS = [
   `angry`
 ];
 
-const COMMENT_TEXTS = [
-  `Interesting setting and a good cast`,
-  `Booooooooooring`,
-  `Very very old. Meh`,
-  `Almost two hours? Seriously?`
-];
-
-const COMMENT_AUTHORS = [
-  `Tim Macoveev`,
-  `John Doe`
-];
-
-const FILM_NAMES = [
-  `The man with the golden arm`,
-  `The great flamarion`,
-  `The dance of life`,
-  `Santa Claus conquers the martians`,
-  `Sagebrush trail`,
-  `Popeye meets sinbad`,
-  `Made for each other`
-];
-
-const FILM_POSTERS = [
-  `./images/posters/made-for-each-other.png`,
-  `./images/posters/popeye-meets-sinbad.png`,
-  `./images/posters/sagebrush-trail.jpg`,
-  `./images/posters/santa-claus-conquers-the-martians.jpg`,
-  `./images/posters/the-dance-of-life.jpg`,
-  `./images/posters/the-great-flamarion.jpg`,
-  `./images/posters/the-man-with-the-golden-arm.jpg`
-];
-
-const FILM_GENRES = [
-  `Musical`,
-  `Horror`,
-  `Fantasy`,
-  `Historical`,
-  `Action`,
-  `Comedy`
-];
-
-const COUNTRYES = [
-  `USA`,
-  `Russia`,
-  `Germany`,
-  `France`,
-  `Brazil`,
-  `Czech`
-];
-
-const PEOPLES = [
-  `Anthony Mann`,
-  `Anne Wigton`,
-  `Heinz Herald`,
-  `Richard Weil`,
-  `Erich von Stroheim`,
-  `Mary Beth Hughes`,
-  `Dan Duryea`
-];
-
-const DESCRIPTION_TEXT = `Lorem ipsum dolor sit amet, consectetur adipiscing elit. 
-Cras aliquet varius magna, non porta ligula feugiat eget. 
-Fusce tristique felis at fermentum pharetra. 
-Aliquam id orci ut lectus varius viverra. 
-Nullam nunc ex, convallis sed finibus eget, sollicitudin eget ante. 
-Phasellus eros mauris, condimentum sed nibh vitae, sodales efficitur ipsum. 
-Sed blandit, eros vel aliquam faucibus, purus ex euismod diam, eu luctus nunc ante ut dui. 
-Sed sed nisi sed augue convallis suscipit in sed felis. 
-Aliquam erat volutpat. Nunc fermentum tortor ac porta dapibus. 
-In rutrum ac purus sit amet tempus.`;
-
 const PROFILE_RATINGS = [`novice`, `fan`, `movie buff`];
 const PROFILE_RATING_VALUES = [0, 10, 20];
 
@@ -101,6 +64,12 @@ const FilterType = {
   FAVORITES: `favorites`
 };
 
+const SortType = {
+  DATE: `date`,
+  RATING: `raiting`,
+  DEFAULT: `default`,
+};
+
 const StatisticsFilter = {
   ALL_TIME: `all-time`,
   TODAY: `today`,
@@ -109,23 +78,29 @@ const StatisticsFilter = {
   YEAR: `year`,
 };
 
-const AUTHORIZATION = `Basic sghjtyuyk5667457DSDSFfdfs`;
+const AUTHORIZATION = `Basic sghjtyuyk5887457DSDSFfdfs`;
 
 const END_POINT = `https://11.ecmascript.pages.academy/cinemaddict`;
 
-export {MONTHS,
+export {
+  Method,
+  RenderPosition,
+  Mode,
+  SHAKE_ANIMATION_TIMEOUT,
+  siteBodyElement,
+  SHOWING_CARDS_COUNT_ON_START,
+  SHOWING_CARDS_COUNT_BY_BUTTON,
+  EXTRA_FILM_CARDS_COUNT,
+  FILTER_ID_PREFIX,
+  HIDDEN_CLASS,
+  ITEM_ACTIVE_CLASS,
+  DefaultData,
+  MONTHS,
   COMMENT_EMOJIS,
-  COMMENT_TEXTS,
-  COMMENT_AUTHORS,
-  FILM_NAMES,
-  FILM_POSTERS,
-  FILM_GENRES,
-  COUNTRYES,
-  PEOPLES,
-  DESCRIPTION_TEXT,
   PROFILE_RATING_VALUES,
   PROFILE_RATINGS,
   FilterType,
+  SortType,
   StatisticsFilter,
   AUTHORIZATION,
   END_POINT
